REFACTOR: NoticeDetailView로 리팩토링
