+++ Projects/Features/KNNotice/Sources/Interfaces/BookmarkFormFactory.swift
@@ -6,9 +6,13 @@
 //
 
 import KNDomain
+import SwiftUI
 import UIKit
 
 @MainActor
 public protocol BookmarkFormFactory {
+    associatedtype ContentView: View
+    
+    func makeSwiftUIView(for notice: Notice, dismissAction: @escaping () -> Void) -> ContentView
     func make(for notice: Notice) -> UIViewController
 }
